requirement.txt
langchain==0.0.180
openai==0.27.7
pandasai==0.2.15
pdfplumber
pypdf==3.9.0
streamlit==1.22.0
streamlit-chat==0.0.2.2
tiktoken==0.3.0
youtube-transcript-api==0.6.0
faiss-cpu
tokenizer
tokenizers 
